feat: Implement PagoFácil integration with transaction management
- Created PagoFacilTransaction model to handle transactions.
- Developed CuotasService for generating and managing payment installments.
- Added PagoFacilService for interacting with PagoFácil API, including authentication, QR generation, and transaction querying.
- Configured PagoFácil settings in a new configuration file.
- Created migration for pagofacil_transactions table to store transaction data.
- Developed PaymentQRModal component for user interaction with payment methods and QR code generation.

+++ database/migrations/2025_11_25_055036_create_detalle_ventas_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id('codigo_detalle_venta');
             $table->foreignId('venta_id')->constrained('ventas', 'codigo_venta')->onDelete('cascade');
             $table->foreignId('producto_id')->constrained('productos', 'codigo_producto');
-            $table->decimal('cantidad', 10, 2);
+            $table->integer('cantidad');
             $table->decimal('precio_unitario', 10, 2);
             $table->decimal('subtotal', 10, 2);
             $table->timestamps();
